Release v0.5.1
This release fixes a bug
If an action were to take place during a `componentWillMount`, this
will cause an infinite loop as the action triggers `combineLatest`
to emit an event, which then triggers another `renderToString`
which starts the whole thing over again.

There is a `first` filter in the stream, but internally the `onNext` of
the observer is called before the `onCompleted` of the observer can be
called, meaning the `first` filter never has a chance to call `onCompleted`

Moral of the story? Beware of Zalgo!